refactor(ui): migrate VirtualList, Dialog, CommandPalette imports from common/ to @/components/ui
Pure import-path migration. No logic changes.
- LogViewer: import VirtualList from @/components/ui
- LogcatPanel: import VirtualList, type VirtualListHandle from @/components/ui
- DeviceSidebar: import showDialog from @/components/ui
- DevicePanel: import showDialog from @/components/ui
- App.tsx: import DialogHost, CommandPalette, openPalette from @/components/ui

TypeScript: 0 errors
Tests: 743 passed
