+++ src/components/device/DevicePanel.tsx
@@ -27,7 +27,7 @@ import type { Device, AvdInfo } from "@/bindings";
 import { toggleDeviceSidebar } from "@/stores/ui.store";
 import { showToast } from "@/components/ui";
 import { Icon } from "@/components/ui";
-import { showDialog } from "@/components/common/Dialog";
+import { showDialog } from "@/components/ui";
 import { CreateDeviceDialog } from "./CreateDeviceDialog";
 
 export interface DevicePanelProps {
