+++ src/components/device/DeviceSidebar.tsx
@@ -32,7 +32,7 @@ import type { Device, AvdInfo } from "@/bindings";
 import { uiState, toggleDeviceSidebar } from "@/stores/ui.store";
 import { showToast } from "@/components/ui";
 import { Icon } from "@/components/ui";
-import { showDialog } from "@/components/common/Dialog";
+import { showDialog } from "@/components/ui";
 import { CreateDeviceDialog } from "./CreateDeviceDialog";
 import { DownloadSystemImageDialog } from "./DownloadSystemImageDialog";
